# backend/src/docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Tours Connect API
  description: |
    Hackathon MVP API for Tours Connect — an AI-powered tourism marketplace.
    Supports customer discovery, saved experiences, booking requests, and provider listings.
    Data is stored in AWS DynamoDB.
  version: 1.0.0
  contact:
    name: Tours Connect
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new customer or provider account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get the authenticated user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/PublicUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/experiences:
    get:
      tags: [Experiences]
      summary: Browse and search published experiences
      parameters:
        - $ref: "#/components/parameters/destination"
        - $ref: "#/components/parameters/category"
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/featured"
        - $ref: "#/components/parameters/minPrice"
        - $ref: "#/components/parameters/maxPrice"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: Paginated list of experiences
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/PaginatedExperiences"
    post:
      tags: [Experiences]
      summary: Create a new experience listing (provider only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExperienceRequest"
      responses:
        "201":
          description: Experience created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Experience"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/experiences/featured:
    get:
      tags: [Experiences]
      summary: Get featured experiences for the homepage
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 20
      responses:
        "200":
          description: Featured experiences
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Experience"
  /api/experiences/{experienceId}:
    get:
      tags: [Experiences]
      summary: Get experience details
      parameters:
        - $ref: "#/components/parameters/experienceId"
      responses:
        "200":
          description: Experience details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Experience"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Experiences]
      summary: Update an experience listing (owner provider or admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/experienceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExperienceRequest"
      responses:
        "200":
          description: Experience updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Experience"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/providers:
    post:
      tags: [Providers]
      summary: Create a provider business profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProviderRequest"
      responses:
        "201":
          description: Provider profile created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Provider"
        "409":
          description: Provider profile already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/providers/me:
    get:
      tags: [Providers]
      summary: Get the authenticated provider profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Provider profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Provider"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/providers/{providerId}:
    get:
      tags: [Providers]
      summary: Get a provider public profile
      parameters:
        - $ref: "#/components/parameters/providerId"
      responses:
        "200":
          description: Provider profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Provider"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/providers/{providerId}/experiences:
    get:
      tags: [Providers]
      summary: List experiences offered by a provider
      parameters:
        - $ref: "#/components/parameters/providerId"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Provider experiences
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/PaginatedExperiences"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /api/auth/login or /api/auth/register
  parameters:
    experienceId:
      name: experienceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    providerId:
      name: providerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    destination:
      name: destination
      in: query
      schema:
        type: string
      example: Lagos
    category:
      name: category
      in: query
      schema:
        $ref: "#/components/schemas/ExperienceCategory"
    search:
      name: search
      in: query
      schema:
        type: string
    featured:
      name: featured
      in: query
      schema:
        type: boolean
    minPrice:
      name: minPrice
      in: query
      schema:
        type: number
    maxPrice:
      name: maxPrice
      in: query
      schema:
        type: number
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 50
    cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Pagination cursor from a previous response
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    UserRole:
      type: string
      enum: [customer, provider, admin]
    ExperienceCategory:
      type: string
      enum:
        - adventure
        - relaxation
        - nightlife
        - cultural
        - wildlife
        - water_sports
        - romantic
        - family_friendly
    ExperienceStatus:
      type: string
      enum: [draft, published, archived]
...
    PublicUser:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        phone:
          type: string
        createdAt:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
          default: customer
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/PublicUser"
            token:
              type: string
              description: JWT bearer token
    Experience:
      type: object
      properties:
        experienceId:
          type: string
          format: uuid
        providerId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        destination:
          type: string
        destinationSlug:
          type: string
        category:
          $ref: "#/components/schemas/ExperienceCategory"
        price:
          type: number
        currency:
          type: string
          example: NGN
        duration:
          type: string
          example: 3 hours
        maxGroupSize:
          type: integer
        images:
          type: array
          items:
            type: string
            format: uri
        featured:
          type: boolean
        status:
          $ref: "#/components/schemas/ExperienceStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateExperienceRequest:
      type: object
      required:
        - title
        - description
        - destination
        - category
        - price
        - duration
        - maxGroupSize
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
          minLength: 10
        destination:
          type: string
        category:
          $ref: "#/components/schemas/ExperienceCategory"
        price:
          type: number
          minimum: 0
        currency:
          type: string
          default: NGN
        duration:
          type: string
        maxGroupSize:
          type: integer
          minimum: 1
        images:
          type: array
          items:
            type: string
            format: uri
        featured:
          type: boolean
          default: false
        status:
          type: string
          enum: [draft, published]
          default: draft
    UpdateExperienceRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        destination:
          type: string
        category:
          $ref: "#/components/schemas/ExperienceCategory"
        price:
          type: number
        currency:
          type: string
        duration:
          type: string
        maxGroupSize:
          type: integer
        images:
          type: array
          items:
            type: string
            format: uri
        featured:
          type: boolean
        status:
          $ref: "#/components/schemas/ExperienceStatus"
    PaginatedExperiences:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Experience"
        nextCursor:
          type: string
...
    Provider:
      type: object
      properties:
        providerId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        businessName:
          type: string
        description:
          type: string
        location:
          type: string
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProviderRequest:
      type: object
      required: [businessName, description, location]
      properties:
        businessName:
          type: string
        description:
          type: string
        location:
          type: string
...